fixed the following: - changed size in info jobs command from int to unsigned long - perform inventory and connector initialization within Server::run - do not remove attributes if migration gets suspended - formatting

+++ src/client/InfoJobsCommand.cc
@@ -77,7 +77,7 @@ void InfoJobsCommand::doCommand(int argc, char **argv)
         std::string filename = infojobsresp.filename();
         recnum = infojobsresp.reqnumber();
         int replnum = infojobsresp.replnumber();
-        int size = infojobsresp.filesize();
+        unsigned long size = infojobsresp.filesize();
         std::string tapeid = infojobsresp.tapeid();
         std::string state = infojobsresp.state();
         if (recnum != Const::UNSET)
